Create separate file for api client
