+++ guides/decoding-encoding/api-client.js
@@ -1,0 +1,12 @@
+const { apiClient } = require('@liskhq/lisk-client');
+let clientCache;
+const nodeAPIURL = 'ws://localhost:7887/rpc-ws';
+
+const getClient = async () => {
+	if (!clientCache) {
+		clientCache = await apiClient.createWSClient(nodeAPIURL);
+	}
+	return clientCache;
+};
+
+module.exports = { getClient };
