+++ guides/decoding-encoding/getBlockById.js
@@ -1,14 +1,5 @@
-const { apiClient } = require('@liskhq/lisk-client');
+const { getClient } = require('./api-client');
 const assert = require("assert");
-let clientCache;
-const nodeAPIURL = 'ws://localhost:7887/rpc-ws';
-
-const getClient = async () => {
-	if (!clientCache) {
-		clientCache = await apiClient.createWSClient(nodeAPIURL);
-	}
-	return clientCache;
-};
 
 if (process.argv.length < 3) {
 	console.error("Please provide the block ID to be decoded.")
